services/analytics-service/requirements.txt
flask==2.3.2
cloudevents==1.10.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.8

